@@ -11,48 +11,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.7",
-    "@angular/common": "~8.2.7",
-    "@angular/compiler": "~8.2.7",
-    "@angular/core": "~8.2.7",
-    "@angular/forms": "~8.2.7",
-    "@angular/platform-browser": "~8.2.7",
-    "@angular/platform-browser-dynamic": "~8.2.7",
-    "@angular/router": "~8.2.7",
-    "@ng-bootstrap/ng-bootstrap": "^5.3.1",
-    "@ngrx/effects": "^10.0.1",
-    "@ngrx/router-store": "^10.0.1",
-    "@ngrx/store": "^10.0.1",
-    "@ngrx/store-devtools": "^10.0.1",
+    "@angular/animations": "~11.2.5",
+    "@angular/common": "~11.2.5",
+    "@angular/compiler": "~11.2.5",
+    "@angular/core": "~11.2.5",
+    "@angular/forms": "~11.2.5",
+    "@angular/localize": "^11.2.5",
+    "@angular/platform-browser": "~11.2.5",
+    "@angular/platform-browser-dynamic": "~11.2.5",
+    "@angular/router": "~11.2.5",
+    "@ng-bootstrap/ng-bootstrap": "^9.0.2",
+    "@ngrx/effects": "^11.0.1",
+    "@ngrx/router-store": "^11.0.1",
+    "@ngrx/store": "^11.0.1",
+    "@ngrx/store-devtools": "^11.0.1",
     "@ngx-translate/core": "^12.1.2",
     "@ngx-translate/http-loader": "^6.0.0",
     "bootstrap": "^4.5.2",
     "jquery": "^3.5.1",
     "material-icons": "^0.3.1",
-    "rxjs": "~6.4.0",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "rxjs": "~6.6.6",
+    "tslib": "^2.1.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.5",
-    "@angular/cli": "~8.3.5",
-    "@angular/compiler-cli": "~8.2.7",
-    "@angular/language-service": "~8.2.7",
-    "@types/jasmine": "~3.3.8",
+    "@angular-devkit/build-angular": "~0.1102.4",
+    "@angular/cli": "~11.2.4",
+    "@angular/compiler-cli": "~11.2.5",
+    "@angular/language-service": "~11.2.5",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.0.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^5.2.3",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "6.0.1",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.2.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "prettier": "^2.1.2",
-    "protractor": "^7.0.0",
+    "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.5.3"
+    "tslint": "~6.1.0",
+    "typescript": "~4.0.7"
   }
 }
